get rid of useSsuQueyLink
